refactor(protonup): split install module
