+++ src/crosshook-native/crates/crosshook-core/src/protonup/install/tests/mod.rs
@@ -1,0 +1,4 @@
+mod archive;
+mod orchestrator;
+mod support;
+mod validation;
